chore(brand): rename package and project from listai to monopop
- app.json: name, slug, scheme, package → monopop; versionCode 20→23
- build.gradle: namespace + applicationId → com.mahjard.monopop
- AndroidManifest: deep link scheme listai:// → monopop://
- strings.xml: app_name Listai → Monopop
- settings.gradle: rootProject.name Listai → Monopop
- package.json: name + version → monopop 1.7.0
- database.ts: default db file listai.db → monopop.db
- Kotlin sources moved to com/mahjard/monopop/
