+++ android/app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         <action android:name="android.intent.action.VIEW"/>
         <category android:name="android.intent.category.DEFAULT"/>
         <category android:name="android.intent.category.BROWSABLE"/>
-        <data android:scheme="listai"/>
+        <data android:scheme="monopop"/>
       </intent-filter>
       <intent-filter>
         <action android:name="android.intent.action.VIEW"/>
