+++ android/settings.gradle
@@ -31,7 +31,7 @@ extensions.configure(com.facebook.react.ReactSettingsExtension) { ex ->
 }
 expoAutolinking.useExpoModules()
 
-rootProject.name = 'Listai'
+rootProject.name = 'Monopop'
 
 expoAutolinking.useExpoVersionCatalog()
