+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "listai",
+  "name": "monopop",
   "author": "Mah",
-  "version": "1.2.0",
+  "version": "1.7.0",
   "license": "Apache-2.0",
   "main": "node_modules/expo/AppEntry.js",
   "engines": {
